Remove hard dependency on cuda driver library
amun/amunmt binaries have a dependency on the CUDA driver library. As far as I can tell, it only uses the runtime API, so the driver dependency can be discarded. This is not a huge problem, but it can cause a compilation error on systems where libcuda.so is placed outside of the normal library search path.