Add Nematus-style factor support
Refactor to use a structured list of factors everywhere

Sentence now contains both a factored representation of the sentence
and a seperate vector with only the word surface forms (a copy of the
first elements of the factor vectors). This allows to not complicate
things in the decoder which doesn't care about factors.

Fill in dummy factors even if there are only words