Allow build dir to be outside the source tree
Currently, the call to git to record the git revision fails if the build
directory is outside the source tree. This fix allows the build directory
to be outside the source tree. However, compilation still works only if
the source is obtained via git and not from a .zip file.