Don't close named pipes during corpus reset.
The corpus reset reopened input files for corpora. So files were opened once during
the construction of the Corpus instance, then closed and repopened during data->reset()
triggered by BatchGenerator::prepare(false). With normal files that's not a problem,
but if the "file" is a named pipe, the closing triggers a SIGPIPE (broken pipe) on
the writing end of the pipe. With this commit, Corpus::reset() leaves open pipes
alone.