Merged PR 9998: new build option NO_BOOST to disable inclusion of Boost and all features that
This PR is to allow Marian to be built on Windows without Boost, since Boost 1.66 causes build errors under Flo.

With this PR, if the preprocessor symbol `NO_BOOST` is defined, Boost will not be included anywhere, and configurations that rely on that cannot be used.

The long-term goal is to remove Boost completely. This `#define` allows to easily locate all relevant locations.

Boost is currently used for two things:

* the timer in the AutoTuner
* when shuffling the corpus to a temp file (not --shuffle-in-ram)

With `NO_BOOST`, attempting to shuffle to file will terminate the program, and auto-tuner will select the first algorithm.

Not tested with Linux.