Merged PR 10205: towards multiple separated parameter types
This is a helper PR for making Young's packed models cleaner. Parameter objects in Marian are meant to be of one type (say float32) and cannot be correctly mixed for various reasons. This introduces a type -> parameter object map that separates parameters with different types into different spaces. Currently this only works reliably for inference (I hope) and will abort for training if more than one parameter type are specified. With this we will be able to have a parameter object for packed16 and one for float32.

It should also be possible in the future to have the same name for a parameter but differentiated by type, which might solve the tied-embedding problem we discussed a while ago.