Merged PR 11312: Guard scheduler against circular references
This is the first PR in a series to get my graph-group* clean-up into master. End-goal is to have clean label-based updates and fp16 in master.

This small change guards against circular references in the scheduler. For large models the freeing order is wrong which can prevent GPU memory to be freed at the correct time leaving no memory for the actual model. The weak references solve that.