Merged PR 15155: Enable output layer without output bias
Adds `--output-omit-bias` option which allows to train an output layer without a bias vector. This is expected to be useful for `--output-approx-knn` during decoding, as the LSH-based k-NN search is then exactly approximating the correct top-K values for decoding. The bias adds a shift otherwise. In first experiments the lack of the output bias does not seem to result in any performance loss.